(#1) Add back the original check as a precaution
There are still cases where users forget to change the local mail
server setting in DirectAdmin. The original check will catch those
mistakes which is why it makes sense to add it back in. If a user
has only changed the A record of an existing MX record and did not
change the local mail server setting it will not find it.